Move CLI code into its own module and add first CLI tests
